Move virtual hooks into their own file & template their setup
