+++ src/cs2_sdk/entity/ctakedamageinfo.h
@@ -106,18 +106,18 @@ public:
 		m_bitsDamageType = bitsDamageType;
 	}
 
-	Vector m_vecDamageForce;			// 0x8  |  8
-	Vector m_vecDamagePosition;			// 0x14 | 20
-	Vector m_vecReportedPosition;		// 0x20 | 32
-	Vector m_vecDamageDirection;		// 0x2c | 44
-	CHandle<CBaseEntity> m_hInflictor;	// 0x38 | 56
-	CHandle<CBaseEntity> m_hAttacker;	// 0x3c | 60
-	CHandle<CBaseEntity> m_hAbility;	// 0x40 | 64
-	float m_flDamage;					// 0x44 | 68
-	float m_flTotalledDamage;			// 0x48 | 72
-	int32_t m_bitsDamageType;			// 0x4c | 76
-	int32_t m_iDamageCustom;			// 0x50 | 80
-	int8_t m_iAmmoType;					// 0x54 | 84
+	Vector m_vecDamageForce;		   // 0x8  |  8
+	Vector m_vecDamagePosition;		   // 0x14 | 20
+	Vector m_vecReportedPosition;	   // 0x20 | 32
+	Vector m_vecDamageDirection;	   // 0x2c | 44
+	CHandle<CBaseEntity> m_hInflictor; // 0x38 | 56
+	CHandle<CBaseEntity> m_hAttacker;  // 0x3c | 60
+	CHandle<CBaseEntity> m_hAbility;   // 0x40 | 64
+	float m_flDamage;				   // 0x44 | 68
+	float m_flTotalledDamage;		   // 0x48 | 72
+	int32_t m_bitsDamageType;		   // 0x4c | 76
+	int32_t m_iDamageCustom;		   // 0x50 | 80
+	int8_t m_iAmmoType;				   // 0x54 | 84
 
 private:
 	[[maybe_unused]] uint8_t m_nUnknown0[0xb]; // 0x55 | 85
